Extract WatchedFolderControl from ImportServiceHandle
The watched-folders / scan-driving identity moves out of the God-ish
ImportServiceHandle into a composed WatchedFolderControl that holds the
watched-folder registry and the folder-watcher command channel, plus
clones of the broadcast sender, runtime handle, and library dir those
methods need. The handle keeps one `watched_folders` field and delegates
its public watched-folder API to it, with every public signature
unchanged so bae-bridge is untouched.

The control never points back at the handle. The only library_manager
touch in the moved methods was `library_dir()` for the registry's save
path, so it holds a LibraryDir directly rather than the whole manager.
