+++ bae-core/src/import/handle/mod.rs
@@ -16,6 +16,9 @@ mod import;
 mod scan;
 mod search;
 mod watch;
+mod watched_folders;
+
+use watched_folders::WatchedFolderControl;
 
 #[cfg(test)]
 mod tests;
@@ -154,12 +157,12 @@ pub struct ImportServiceHandle {
     pub progress_handle: ImportProgressHandle,
     pub library_manager: LibraryManager,
     pub runtime_handle: tokio::runtime::Handle,
-    pub watcher_tx: mpsc::UnboundedSender<WatcherCommand>,
     /// Unified event channel — all import service events go here.
     pub event_tx: broadcast::Sender<ImportEvent>,
-    /// The persistent watched-folder list. Mutated by `add_watched_folder` /
-    /// `remove_watched_folder`, which persist it and broadcast the new list.
-    pub folder_registry: Arc<Mutex<ImportFolderRegistry>>,
+    /// The watched-folders / scan-driving surface: the watched-folder list, the
+    /// folder-watcher command channel, and the shared handles those methods
+    /// need. The public watched-folder methods delegate here.
+    watched_folders: WatchedFolderControl,
     cover_art_archive: CoverArtArchiveClient,
 }
 
@@ -212,14 +215,20 @@ impl ImportServiceHandle {
     ) -> Self {
         let progress_handle =
             ImportProgressHandle::new(event_tx.subscribe(), runtime_handle.clone());
+        let watched_folders = WatchedFolderControl::new(
+            folder_registry,
+            watcher_tx,
+            event_tx.clone(),
+            runtime_handle.clone(),
+            library_manager.library_dir().clone(),
+        );
         Self {
             requests_tx,
             progress_handle,
             library_manager,
             runtime_handle,
-            watcher_tx,
             event_tx,
-            folder_registry,
+            watched_folders,
             cover_art_archive,
         }
     }
